refactor(api): migrate Spotify API service to RTK Query
- Replace manual SpotifyApiService class with RTK Query API slice
- Move API implementation from src/features/api to src/services
- Define three endpoints with tag-based caching: getArtist, getTrack, getAudioFeatures
- Update store configuration to use RTK Query middleware
- Remove manual error handling and validation (now handled by RTK Query)
- Update task definitions to reflect new file structure
