+++ src/services/index.ts
@@ -1,0 +1,17 @@
+/**
+ * Services Exports
+ *
+ * Central export point for all API services and related hooks.
+ * This ensures consistent importing patterns across the application.
+ *
+ * Usage:
+ *   import { spotifyApi, useGetArtistQuery } from '@/services'
+ */
+
+export {
+  spotifyApi,
+  useGetArtistQuery,
+  useGetTrackQuery,
+  useGetAudioFeaturesQuery,
+} from './spotify-api';
+export { default } from './spotify-api';
